# src/routing/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IErrorResponseDto": {
				"description": "Тело любого ответа с ошибкой (`status >= 400`).",
				"properties": {
					"status": {
						"type": "number",
						"format": "double",
						"description": "HTTP-статус."
					},
					"code": {
						"type": "string",
						"description": "Машинный код: доменный (`USER_EMAIL_TAKEN`) или по статусу (`NOT_FOUND`,\n`VALIDATION_ERROR`, `INTERNAL_ERROR`). Стабилен — на него опирается клиент."
					},
					"message": {
						"type": "string",
						"description": "Человекочитаемое сообщение; для 5xx — общее, без внутренних деталей."
					},
					"details": {
						"description": "Подробности: для VALIDATION_ERROR — поле → сообщение."
					},
					"requestId": {
						"type": "string",
						"description": "Идентификатор запроса — для обращения в поддержку и поиска в логах."
					},
					"stack": {
						"type": "string",
						"description": "Stack trace — только вне production."
					}
				},
				"required": [
					"status",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UUID": {
				"type": "string",
				"description": "Идентификатор сущности в пути или теле запроса. tsoa проверяет формат до\nконтроллера: неверный uuid → 422, а не ошибка Postgres.",
				"pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
			},
			"TPermission": {
				"type": "string",
				"description": "Право — строка `domain:action` (или `domain:*` — wildcard). Модуль объявляет\nсвои права сам через `definePermissions`; каталог с подписями отдаёт\n`GET /api/v1/permissions`."
			},
...
			"KnownRole": {
				"type": "string",
				"enum": [
					"admin",
					"user",
					"guest"
				],
				"description": "Тип для предопределённых ролей (автодополнение в IDE)."
			},
			"TRole": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/KnownRole"
					},
					{
						"allOf": [
							{
								"type": "string"
							},
							{
								"properties": {},
								"type": "object"
							}
						]
					}
				],
				"description": "Роль — произвольная строка; предопределённые значения дают автодополнение."
			},
			"IPermissionDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/TPermission"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRoleDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/TRole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RoleName": {
				"type": "string",
				"description": "Имя роли в теле запроса API. Обычная строка: `TRole` с `string & {}`\ntsoa проверить не может и отклоняет любое значение. Известные роли — `Roles`.",
				"minLength": 1,
				"maxLength": 100
			},
			"ICreateRoleRequestDto": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/RoleName"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PermissionName": {
				"type": "string",
				"description": "Имя права в теле запроса API (`домен:действие`).",
				"minLength": 1,
				"maxLength": 100
			},
			"IRolePermissionsRequestDto": {
				"properties": {
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/PermissionName"
						},
						"type": "array"
					}
				},
				"required": [
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"apiKey": {
				"type": "apiKey",
				"in": "header",
				"name": "X-Api-Key",
				"description": "API-ключ: агент ноды (scope wg-agent) или интеграция"
			}
		}
	},
	"info": {
		"title": "wg-admin",
		"version": "1.0.1",
		"description": "WG Admin — админка WireGuard: бэкенд (Koa, tsoa, TypeORM, pg-boss) и агент нод на Go",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"paths": {
...
		"/api/v1/roles": {
			"get": {
				"operationId": "GetRoles",
				"responses": {
					"200": {
						"description": "Список ролей",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IRoleDto"
									},
									"type": "array"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Получить все роли с их правами.",
				"summary": "Список ролей",
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": [
							"permission:role:view"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateRole",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRoleDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Создать новую роль.",
				"summary": "Создание роли",
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": [
							"permission:role:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Название роли",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateRoleRequestDto",
								"description": "Название роли"
							}
						}
					}
				}
			}
		},
		"/api/v1/roles/{id}": {
			"delete": {
				"operationId": "DeleteRole",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Удалить роль. Системные роли (`admin`, `user`, `guest`) не удаляются,\nсобственную роль удаляет только суперпользователь.",
				"summary": "Удаление роли",
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": [
							"permission:role:delete"
						]
					}
				],
				"parameters": [
					{
						"description": "ID роли",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/api/v1/roles/{id}/permissions": {
			"patch": {
				"operationId": "SetRolePermissions",
				"responses": {
					"200": {
						"description": "Обновлённая роль",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRoleDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Установить права для роли.\nЗаменяет текущий набор прав роли указанным. Роль `admin`, право `*` и\nсобственную роль меняет только суперпользователь. Все пользователи роли\nполучают `user:privileges-changed` с новыми правами; их прежние\naccess-токены отклоняются (`AUTH_PRIVILEGES_CHANGED`), сессии остаются.",
				"summary": "Установка прав роли",
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": [
							"permission:role:update"
						]
					}
				],
				"parameters": [
					{
						"description": "ID роли",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"description": "Список прав",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRolePermissionsRequestDto",
								"description": "Список прав"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		}
	]
}